Login should now persist.
Login should now persist. If you create a new route, just bring in the SQL_CRED and use it to create a connection (con) > run the function you need to in the dataconnector the inside a  [with x as y] and the connection will open and close on its own. There are 2 examples in the dataconnector.py.